requirements.txt
#Python3.11.13
django==5.2.4 
mysqlclient==2.2.7
django-admin-interface==0.30.1
uvicorn
djangorestframework
python-dotenv